Only use new boot linkup server function
