+++ linkup-cli/src/background_booting.rs
@@ -12,16 +12,13 @@ use url::Url;
 
 use crate::local_config::{LocalState, ServiceTarget};
 use crate::services::local_server::{is_local_server_started, start_local_server};
-use crate::services::tunnel::{CfTunnelManager, TunnelManager};
-use crate::status::print_session_names;
 use crate::worker_client::WorkerClient;
-use crate::{linkup_file_path, services, LINKUP_LOCALSERVER_PORT};
-use crate::{CliError, LINKUP_LOCALDNS_INSTALL};
+use crate::CliError;
+use crate::{services, LINKUP_LOCALSERVER_PORT};
 
 #[cfg_attr(test, mockall::automock)]
 pub trait BackgroundServices {
     fn boot_linkup_server(&self, state: LocalState) -> Result<LocalState, CliError>;
-    fn boot_background_services(&self, state: LocalState) -> Result<LocalState, CliError>;
     fn boot_local_dns(&self, domains: Vec<String>, session_name: String) -> Result<(), CliError>;
 }
 
@@ -66,46 +63,6 @@ impl BackgroundServices for LocalBackgroundServices {
         services::dnsmasq::start(domains, session_name)?;
 
         Ok(())
-    }
-
-    fn boot_background_services(&self, mut state: LocalState) -> Result<LocalState, CliError> {
-        state = self.boot_linkup_server(state)?;
-
-        let should_run_free = state.linkup.is_paid.is_none() || !state.linkup.is_paid.unwrap();
-        if should_run_free {
-            if state.should_use_tunnel() {
-                let tunnel_manager = CfTunnelManager {};
-                if tunnel_manager.is_tunnel_running().is_err() {
-                    println!("Starting tunnel...");
-                    let tunnel = tunnel_manager.run_tunnel(&state)?;
-                    state.linkup.tunnel = Some(tunnel);
-                } else {
-                    println!("Cloudflare tunnel was already running.. Try stopping linkup first if you have problems.");
-                }
-            } else {
-                println!(
-                "Skipping tunnel start... WARNING: not all kinds of requests will work in this mode."
-            );
-            }
-        }
-
-        if should_run_free {
-            if linkup_file_path(LINKUP_LOCALDNS_INSTALL).exists() {
-                self.boot_local_dns(state.domain_strings(), state.linkup.session_name.clone())?;
-            }
-
-            if let Some(tunnel) = &state.linkup.tunnel {
-                println!("Waiting for tunnel DNS to propagate at {}...", tunnel);
-
-                wait_for_dns_ok(tunnel.clone())?;
-
-                println!();
-            }
-        }
-
-        print_session_names(&state);
-
-        Ok(state)
     }
 }
